requirements_gpu.txt
soundfile>=0.9.0
tensorflow-gpu>=1.3.0
pyyaml>=3.12